Fix: validate that attributes are non-empty where we expect them to be
This shows a ton of pages that have attributes that simply do nothing.
One could argue we should silently ignore those, but the user might
think it does have an effect. As such, it is better to decline these
pages to render properly, in the assumption someone will fix it.